+++ app/app_test.go
@@ -3,8 +3,8 @@ package app_test
 import (
 	"net/netip"
 
-	"github.com/lysShub/itun/crypto"
-	"github.com/lysShub/itun/sconn"
+	"github.com/lysShub/fatun/crypto"
+	"github.com/lysShub/fatun/sconn"
 	"gvisor.dev/gvisor/pkg/tcpip/header"
 )
