+++ app/app_windows_test.go
@@ -13,11 +13,11 @@ import (
 	_ "net/http/pprof"
 
 	"github.com/lysShub/divert-go"
-	"github.com/lysShub/itun/app"
-	"github.com/lysShub/itun/app/client"
-	"github.com/lysShub/itun/app/client/capture"
-	"github.com/lysShub/itun/app/client/filter"
-	"github.com/lysShub/itun/sconn"
+	"github.com/lysShub/fatun/app"
+	"github.com/lysShub/fatun/app/client"
+	"github.com/lysShub/fatun/app/client/capture"
+	"github.com/lysShub/fatun/app/client/filter"
+	"github.com/lysShub/fatun/sconn"
 	"github.com/lysShub/sockit/conn/tcp"
 	dconn "github.com/lysShub/sockit/conn/tcp/divert"
 	"github.com/lysShub/sockit/packet"
