+++ app/client/client.go
@@ -7,12 +7,12 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/lysShub/itun/app"
-	"github.com/lysShub/itun/app/client/capture"
-	cs "github.com/lysShub/itun/app/client/session"
-	"github.com/lysShub/itun/control"
-	"github.com/lysShub/itun/sconn"
-	"github.com/lysShub/itun/session"
+	"github.com/lysShub/fatun/app"
+	"github.com/lysShub/fatun/app/client/capture"
+	cs "github.com/lysShub/fatun/app/client/session"
+	"github.com/lysShub/fatun/control"
+	"github.com/lysShub/fatun/sconn"
+	"github.com/lysShub/fatun/session"
 	"github.com/lysShub/sockit/errorx"
 	"github.com/lysShub/sockit/packet"
 	"gvisor.dev/gvisor/pkg/tcpip/header"
