+++ app/server/proxyer/proxyer_impl.go
@@ -6,10 +6,10 @@ import (
 	"net/netip"
 	"time"
 
-	"github.com/lysShub/itun/app/server/adapter"
-	ss "github.com/lysShub/itun/app/server/proxyer/session"
-	"github.com/lysShub/itun/control"
-	"github.com/lysShub/itun/session"
+	"github.com/lysShub/fatun/app/server/adapter"
+	ss "github.com/lysShub/fatun/app/server/proxyer/session"
+	"github.com/lysShub/fatun/control"
+	"github.com/lysShub/fatun/session"
 	"github.com/lysShub/sockit/errorx"
 	"github.com/lysShub/sockit/packet"
 )
