+++ cmd/client/main_windows.go
@@ -11,12 +11,12 @@ import (
 	"os"
 
 	"github.com/lysShub/divert-go"
-	"github.com/lysShub/itun/app"
-	"github.com/lysShub/itun/app/client"
-	"github.com/lysShub/itun/app/client/capture"
-	"github.com/lysShub/itun/app/client/filter"
-	"github.com/lysShub/itun/crypto"
-	"github.com/lysShub/itun/sconn"
+	"github.com/lysShub/fatun/app"
+	"github.com/lysShub/fatun/app/client"
+	"github.com/lysShub/fatun/app/client/capture"
+	"github.com/lysShub/fatun/app/client/filter"
+	"github.com/lysShub/fatun/crypto"
+	"github.com/lysShub/fatun/sconn"
 	"github.com/lysShub/sockit/conn/tcp"
 	dconn "github.com/lysShub/sockit/conn/tcp/divert"
 	"github.com/lysShub/sockit/test"
