+++ faketcp/faketcp.go
@@ -4,7 +4,7 @@ import (
 	"math/rand"
 	"sync/atomic"
 
-	"github.com/lysShub/itun/crypto"
+	"github.com/lysShub/fatun/crypto"
 	"github.com/lysShub/sockit/errorx"
 	"github.com/lysShub/sockit/packet"
 	"github.com/stretchr/testify/require"
