+++ sconn/handshake.go
@@ -6,11 +6,11 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/lysShub/itun/crypto"
-	"github.com/lysShub/itun/faketcp"
-	"github.com/lysShub/itun/session"
-	"github.com/lysShub/itun/ustack"
-	"github.com/lysShub/itun/ustack/gonet"
+	"github.com/lysShub/fatun/crypto"
+	"github.com/lysShub/fatun/faketcp"
+	"github.com/lysShub/fatun/session"
+	"github.com/lysShub/fatun/ustack"
+	"github.com/lysShub/fatun/ustack/gonet"
 	"github.com/lysShub/sockit/packet"
 	"github.com/lysShub/sockit/test"
 	"github.com/lysShub/sockit/test/debug"
