+++ sconn/listen.go
@@ -6,9 +6,9 @@ import (
 	"net/netip"
 	"sync/atomic"
 
-	"github.com/lysShub/itun/ustack"
-	"github.com/lysShub/itun/ustack/gonet"
-	"github.com/lysShub/itun/ustack/link"
+	"github.com/lysShub/fatun/ustack"
+	"github.com/lysShub/fatun/ustack/gonet"
+	"github.com/lysShub/fatun/ustack/link"
 	"github.com/lysShub/sockit/conn"
 	"github.com/pkg/errors"
