+++ sconn/sconn.go
@@ -9,10 +9,10 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/lysShub/itun/faketcp"
-	"github.com/lysShub/itun/session"
-	"github.com/lysShub/itun/ustack"
-	"github.com/lysShub/itun/ustack/gonet"
+	"github.com/lysShub/fatun/faketcp"
+	"github.com/lysShub/fatun/session"
+	"github.com/lysShub/fatun/ustack"
+	"github.com/lysShub/fatun/ustack/gonet"
 	"github.com/lysShub/sockit/conn"
 	"github.com/lysShub/sockit/errorx"
 	"github.com/lysShub/sockit/packet"
