+++ ustack/ustack.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/lysShub/itun/ustack/link"
+	"github.com/lysShub/fatun/ustack/link"
 	"github.com/lysShub/sockit/helper/ipstack"
 	"github.com/lysShub/sockit/packet"
